refactoring: split up bootstrap_sql_test.py
